Moviendo una funcion de jwt utils al api e importandola correctamente al consumer
